First draft at solving kata 5.
I'm pretty sure this works (and sporn is just a false positive),
though it could use some more testing.  I'd like to make it
faster, though.  Making a bloom filter out of 98K entries in
/usr/share/dict/words takes about a minute on this VM.

My current hypothesis is that a set is a bad datastructure to use.
The other obvious thing is a vector of bools/(1|0)s, but a good
bloom filter is going to be so sparse that that seems wasteful.
Given sufficient time and cleverness, I could probably do
something useful by using a vector of 32-bit ints to represent
parts of the bitmap.  That way, if I have 16-bit hashes (and
thus need 2^16 bits in my filter), I only have to store
2^16-2^5=~2K ints.