Reimplementing the bloom filter with a bit vector.
The idea was that a vector of ints would be substantially
more efficient than a set of ints.  In hindsight, this
would only make sense if bloom filter was substantially
filled, which it shouldn't be to have good accuracy.  So
a little intuition should have shown that this wouldn't
actually help all that much.  Le sigh.

It's also significantly harder to implement, and the code
is a lot less friendly.  So that's not a win.

Next hypothesis is that the problem is how I have to make
a copy of the entire bloom filter data structure every
time I add a word to it.  Maybe zippers would help? Maybe.